fix(torrent): improve get_files() fields handling
This commit fixes the following issues with `Torrent.get_files()`:

If Client.get_torrent() was called with "files" not in `arguments`, then
`get_files()` returns an empty list. This is surprising - it should
raise an error instead.

If Client.get_torrent() was called with arguments=["files"], get_files()
fails because it expects to find `properties` and `wanted`, which were
not requested. For efficiency we should not force users to request
fields they don't need, so leave these unset in this case.