Improve error handling of empty torrent metadata
If an empty torrent file (as in, zero bytes) is passed to
`add_torrent()`, the previous code would fall over and attempt to
awkwardly pass the path itself as the torrent data, resulting in a very
cryptic error being raised. This commit ensures we fail early and
cleanly in this case.